analysis/requirements-conda.txt
scikit-learn
python-duckdb
joblib